Move URLImporter out into it's own component
