+++ src/renderer/style.scss
@@ -327,6 +327,15 @@ $sceneDetailOptionsWidth: 220px;
   }
 }
 
+.URLImporter {
+  & > .u-button {
+    float: right;
+    margin-top: $smallPadding;
+    margin-right: $smallPadding;
+    padding: $smallPadding;
+  }
+}
+
 .LoadingIndicator {
   position: fixed;
   top: 0; right: 0; bottom: 0; left: 0;
